start a Split Phase (154) by applying Extract Function (106) to the code that makes up the second phase
I start a Split Phase (154) by applying Extract Function (106) to the code that makes up the second phase. In this case, that’s the statement printing code, which is in fact the entire content of statement. This, together with all the nested functions, goes into its own toplevel function which I call renderPlainText.