use Change Function Declaration (124) to pass the performance’s play into the calculator.
So far, this new object isn’t doing anything. I want to move behavior into it—and I’d like to start with the simplest thing to move, which is the play record. Strictly, I don’t need to do this, as it’s not varying polymorphically, but this way I’ll keep all the data transforms in one place, and that consistency will make the code clearer. To make this work, I will use Change Function Declaration (124) to pass the
performance’s play into the calculator.